spawn/0,1,2: force distributed mode
Since spawn waits for the remote erlang node to come up, the process
will block forever if distribution wasn't enabled.

Helpfully put the node into distributed mode if spawn is called. If the VM
can't find the long name, the process will crash (this can happen if the
IP address can't be reverse resolved). In that case, the node will need
to be started in distributed mode.

If the caller does not want the node to be part of distribution, they
can call create and pass in the erl command line.